+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Chart component: format numbers and prices using store currency settings.
 - Make `href`/linking optional in SummaryNumber.
 - Fix SummaryNumber example code.
+- New `<Search>` prop named `showClearButton`, that will display a 'Clear' button when the search box contains one or more tags.
 
 # 1.4.0
 - Add download log ip address autocompleter to search component